test(conventional): move verify to commit module add add tests
